requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12.0
psycopg2-binary>=2.9.0
django-environ
drf-yasg

